Add self-update for West and the manifest
The self-update mechanism checks the West and manifest repositories in
west/ for upstream changes and rebases them to the latest version if any
are found.

'west fetch' and 'west pull' run a self-update before fetching any
project data. It probably makes sense to always work with the latest
version of West and the manifest there. The self-update can be
suppressed with --no-selfupdate.

There's also a 'west selfupdate' command for manually updating West
and/or the manifest.

West automatically restarts itself after a self-update, before running
any commands or parsing the manifest. This makes it safe to add a new
feature to West and immediately start using it in the manifest.

The restarting is done with a plain exec(3). Getting this to work
required a few changes:

 - West now uses absolute instead of relative imports, to make it
   possible to execute src/west/main.py directly. This also makes it a
   bit simpler to work on West. Relative imports seem to be discouraged
   when googling around a bit.

   __init__.py was removed from src/west. West is more of an application
   than a package now (though it uses some packages internally).

 - The internal 'cmd' package was renamed to 'commands', to avoid
   clashing with a module in the Python standard library.

   'runner' was renamed to 'runners' as well, for consistency.

Using exec(3) also allowed the argv handling to be simplified a bit.

The tests now have to be run from the West repository root with
'PYTHONPATH=src/west pytest'. Maybe this could be improved. The West
module and packages have to get into sys.path somehow at least.

__init__.py was removed from the test directories as well. Don't think
having the tests in a package helps now.

Four unrelated changes are included as well:

 - args.zephyr_base is no longer set to ZEPHYR_BASE, just the other way
   around. args.zephyr_base is currently unused except for setting
   ZEPHYR_BASE. I think it might be cleaner to add a new variable
   instead of directly modifying 'args' if the ZEPHYR_BASE value is
   needed elsewhere later.

 - WEST_TOPDIR was renamed to WEST_MARKER. The old name made me think
   that it was the path to the directory containing west/.

 - Project-related command-line arguments that are only used by a single
   command are now defined together with the command that uses them.

 - A bug was fixed in the bootstrapper's find_west_topdir(). The 'start'
   argument was being ignored.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>